Begin devel on indel finder

+++ src/large_indel_finder.h
@@ -18,7 +18,7 @@ public:
                          std::vector< std::vector< int > > > > &nucleotide_counts);
 
 private:
-    std::vector< std::pair< int, int > > _determineRanges(const std::vector< std::vector< int > > &nucl);
+    std::vector< std::pair< long, long > > _determineRanges(const std::vector< std::vector< int > > &nucl);
 
     Args& _args;
     std::unordered_map< std::string, std::pair< long, std::vector< std::string > > > _large_indels;
